chore: seperate list component
